requirements.txt
asgiref==3.3.1
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.5
Django==3.1.7
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
google-trans-new==1.1.9
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
murmurhash==1.0.5
numpy==1.20.1
openapi-codec==1.3.2
packaging==20.9
pathy==0.4.0
preshed==3.0.5
pydantic==1.7.3
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
smart-open==3.0.0
spacy==3.0.3
spacy-legacy==3.0.1
sqlparse==0.4.1
srsly==2.4.0
thinc==8.0.1
tqdm==4.58.0
typer==0.3.2
uritemplate==3.0.1
urllib3==1.26.3
wasabi==0.8.2
